// plugins/tosios/adapter/vendor-dist/server/src/states/GameState.d.ts
import { ArraySchema, MapSchema, Schema } from '@colyseus/schema';
import { Bullet, Game, Monster, Player, Prop } from '../entities';
import { Models, Types } from '@tosios/common';
export declare class GameState extends Schema {
    game: Game;
    players: MapSchema<Player>;
    monsters: MapSchema<Monster>;
    props: ArraySchema<Prop>;
    bullets: ArraySchema<Bullet>;
    private map;
    private walls;
    private spawners;
    private actions;
    private onMessage;
    constructor(roomName: string, mapName: string, maxPlayers: number, mode: Types.GameMode, onMessage: (message: Models.MessageJSON) => void);
    update(): void;
    private updateGame;
    private updatePlayers;
    private updateMonsters;
    private updateBullets;
    private handleWaitingStart;
    private handleLobbyStart;
    private handleGameStart;
    private handleGameEnd;
    initializeMap: (mapName: string) => void;
    playerAdd(id: string, name: string): void;
    playerPushAction(action: Models.ActionJSON): void;
    private playerMove;
    private playerRotate;
    private playerShoot;
    private playerUpdateKills;
    playerRemove(id: string): void;
    private setPlayersActive;
    private setPlayersPositionRandomly;
    private getPositionRandomly;
    private setPlayersTeamsRandomly;
    private getSpawnerRandomly;
    private monstersAdd;
    private monsterUpdate;
    private monsterRemove;
    private monstersClear;
    private bulletUpdate;
    private propsAdd;
    private propsClear;
}